Use 32bit integer for loop indices in offload region.

+++ src/QMCHamiltonians/CoulombPBCAA.cpp
@@ -704,12 +704,12 @@ std::vector<CoulombPBCAA::Return_t> CoulombPBCAA::mw_evalSR_offload(const RefVec
       ScopedTimer offload_scope(caa_leader.offload_timer_);
 
       PRAGMA_OFFLOAD("omp target teams distribute num_teams(nw)")
-      for (size_t iw = 0; iw < nw; iw++)
+      for (uint32_t iw = 0; iw < nw; iw++)
       {
         mRealType SR = 0.0;
         PRAGMA_OFFLOAD("omp parallel for reduction(+ : SR)")
-        for (size_t jcol = 0; jcol < total_num; jcol++)
-          for (size_t irow = first; irow < last; irow++)
+        for (uint32_t jcol = 0; jcol < total_num; jcol++)
+          for (uint32_t irow = first; irow < last; irow++)
           {
             const RealType dist = mw_dist[num_padded * (irow - first + iw * this_chunk_size) + jcol];
             if (irow == jcol || (irow * 2 + 1 == total_num && jcol > irow))
